requested changes are made
