+++ components/Asyncapi3Comparison/Asyncapi3SchemaFormatComparison.tsx
@@ -1,10 +1,12 @@
 import React, { useState } from 'react';
 
-import type { Asyncapi3SchemaFormatComparisonProps } from '@/types/Asyncapi3Comparison/Asyncapi3SchemaFormatComparison';
+export interface Asyncapi3SchemaFormatComparisonProps {
+  className?: string;
+}
 
 /**
+ * @param {string} [props.className=''] - Additional CSS classes for styling.
  * @description React component for comparing AsyncAPI schema formats between versions 2.x and 3.0.
- * @param {Asyncapi3SchemaFormatComparisonProps} props - The props for the component.
  */
 export default function Asyncapi3SchemaFormatComparison({ className = '' } : Asyncapi3SchemaFormatComparisonProps) {
   const [hoverState, setHoverState] = useState({
