--- types/Asyncapi3Comparison/AsyncAPI3ChannelComparison.ts
@@ -1,10 +1,0 @@
-export interface HoverState {
-    Paths: boolean;
-    PathItem: boolean;
-    Operation: boolean;
-    Message: boolean;
-}
-
-export interface AsyncAPI3ChannelComparisonProps {
-    className?: string;
-}
