--- types/Asyncapi3Comparison/Asyncapi3IdAndAddressComparison.ts
@@ -1,8 +1,0 @@
-export interface HoverState {
-    Paths: boolean;
-    PathItem: boolean;
-}
-
-export interface AsyncAPI3IdAndAddressComparisonProps {
-    className?: string;
-}
