--- types/Asyncapi3Comparison/Asyncapi3MetaComparison.ts
@@ -1,9 +1,0 @@
-export interface Asyncapi3MetaComparisonProps {
-    className?: string;
-}
-
-export interface HoverState {
-    Info: boolean;
-    Tags: boolean;
-    External: boolean;
-}
