--- types/Asyncapi3Comparison/Asyncapi3OperationComparison.ts
@@ -1,3 +1,0 @@
-export interface Asyncapi3OperationComparisonProps {
-    className?: string;
-  }
