--- types/Asyncapi3Comparison/Asyncapi3ParameterComparison.ts
@@ -1,3 +1,0 @@
-export interface Asyncapi3ParameterComparisonProps {
-    className?: string;
-  }
