--- types/Asyncapi3Comparison/Asyncapi3SchemaFormatComparison.ts
@@ -1,3 +1,0 @@
-export interface Asyncapi3SchemaFormatComparisonProps {
-    className?: string;
-  }
