--- types/Asyncapi3Comparison/Asyncapi3ServerComparison.ts
@@ -1,8 +1,0 @@
-export interface Asyncapi3ServerComparisonProps {
-    className?: string;
-}
-
-export interface HoverState {
-    Host: boolean;
-    Path: boolean;
-}
